+++ ShieldMeBruhReforged/Patches/Player_Patches.cs
@@ -1,16 +1,16 @@
 ﻿using HarmonyLib;
-using ShieldMeBruhReforged.Features;
 
 namespace ShieldMeBruhReforged.Patches;
 
 public static class Player_Patches
 {
     [HarmonyPatch(typeof(Player), nameof(Player.SetLocalPlayer))]
-    private static class HumanoidEquipItemPatch
+    private static class SetLocalPlayerPatch
     {
         static void Postfix(Player __instance)
         {
-            AutoShield.ResetEvent.PerformReset(__instance);
+            if (__instance == Player.m_localPlayer)
+                ShieldMeBruhReforged.AutoShield.ResetPlayerContext();
         }
     }